Rename script for tags
Rename since it encompasses industries in addition to skills and
coursework.
